Attempt 2
This actually is NOT my latest working code.  Mine seems to work on my own device but the second I try to download my work from CoLab to reupload it wigs out.  This is a last ditch effort to see if maybe I can at least get the points because I don't get what's going on with the space install.